requirements.txt
groq-cli
openai
python-dotenv
python-dateutil
pydantic
pytest
requests